backend/requirements.txt
# Core UI and App
fastapi
uvicorn

# LangChain + OpenAI
langchain
langchain-openai
openai

# Tooling and HTTP
requests
python-dotenv

# Payment
paypal-agent-toolkit
paypalrestsdk
stripe
pydantic
pydantic[email]